チャージ完璧

+++ LeftHumanWave.cpp
@@ -51,7 +51,7 @@ void LeftHumanWave::Final() {
 void LeftHumanWave::SetWave(LeftHand* lefthand) {
 	int action = lefthand->GetAction();
 	int pat = lefthand->GetPat();
-	int HP = lefthand->GetHP();
+	float HP = lefthand->GetHP();
 	XMFLOAT3 pos = lefthand->GetPosition();
 	//ÕŒ‚”g‚ÌˆÊ’u‚ðŒˆ‚ß‚é
 	if (action == 0 && pos.y == 0.0f && pat == 5 && !expand && HP > 0) {
